Add some crypto primitives
