+++ iroha2/sys/FixedPoint/__init__.py
