+++ iroha2/sys/i64/__init__.py
@@ -1,0 +1,1 @@
+from ..rust import Enum, Struct, Tuple, Dict
