+++ iroha2/sys/iroha_data_model/fixed/__init__.py
@@ -1,1 +1,3 @@
 from ...rust import Enum, Struct, Tuple, Dict
+
+Fixed = Tuple[str]
