--- iroha2/sys/iroha_schema/iroha_data_model/fixed/__init__.py
@@ -1,1 +1,0 @@
-from ....rust import Enum, Struct, Tuple, Dict
