+++ src/lib.rs
@@ -6,127 +6,18 @@
 use std::ops::{Deref, DerefMut};
 
 use iroha_client::{client, config::Configuration};
-use iroha_crypto::{Hash, PrivateKey, PublicKey};
+use iroha_crypto::{Hash, KeyGenConfiguration, Signature};
+use iroha_crypto::{PrivateKey, PublicKey};
 use iroha_data_model::prelude::*;
 use pyo3::class::basic::PyObjectProtocol;
 use pyo3::class::iter::IterNextOutput;
-use pyo3::exceptions::PyException;
 use pyo3::prelude::*;
 use pyo3::PyIterProtocol;
-use pythonize::PythonizeTypes;
 
+use crate::python::*;
+
+mod python;
 mod types;
-
-macro_rules! wrap_class {
-    (
-        $(
-            $ty:ident {
-                $field:ident : $outer_ty:ty
-            } : $( $derive:ident $(+)? )*
-        ),*
-        $(,)?
-    ) => {$(
-        #[pyclass]
-        #[derive($($derive,)*)]
-        pub struct $ty {
-            $field: $outer_ty,
-        }
-
-        impl Deref for $ty {
-            type Target = $outer_ty;
-            fn deref(&self) -> &Self::Target {
-                &self.$field
-            }
-        }
-
-        impl DerefMut for $ty {
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut self.$field
-            }
-        }
-
-        impl From<$outer_ty> for $ty {
-            fn from(outer: $outer_ty) -> Self {
-                Self {
-                    $field: outer,
-                }
-            }
-        }
-
-        impl From<$ty> for $outer_ty {
-            fn from(from: $ty) -> Self {
-                from.$field
-            }
-        }
-
-        #[pyproto]
-        impl PyObjectProtocol for $ty {
-            fn __str__(&self) -> String {
-                format!("{:#?}", self)
-            }
-        }
-    )*
-        fn register_wrapped_classes(m: &PyModule) -> PyResult<()> {
-            $(m.add_class::<$ty>()?;)*
-            Ok(())
-        }
-    };
-}
-
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub struct Dict<T>(T);
-
-impl<T> Deref for Dict<T> {
-    type Target = T;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl<T> DerefMut for Dict<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl<T> Dict<T> {
-    #[allow(clippy::missing_const_for_fn)]
-    pub fn into_inner(self) -> T {
-        self.0
-    }
-}
-
-struct Pythonizer;
-impl<'py> PythonizeTypes<'py> for Pythonizer {
-    type Dict = types::dict::PythonizeDict<'py>;
-    type List = types::list::PythonizeList<'py>;
-}
-
-impl<'source, T: serde::de::DeserializeOwned> FromPyObject<'source> for Dict<T> {
-    fn extract(obj: &'source PyAny) -> PyResult<Self> {
-        let obj = if obj.hasattr("to_rust")? {
-            obj.call_method0("to_rust")?
-        } else {
-            obj
-        };
-
-        pythonize::depythonize_custom::<Pythonizer, _>(obj)
-            .map_err(to_py_err)
-            .map(Self)
-    }
-}
-
-impl<'source, T: serde::Serialize> IntoPy<PyObject> for Dict<T> {
-    fn into_py(self, py: Python) -> PyObject {
-        #[allow(clippy::expect_used)]
-        pythonize::pythonize_custom::<Pythonizer, _>(py, &self.into_inner())
-            .expect("Lets hope pythonize won't complain :(")
-    }
-}
-
-fn to_py_err(err: impl Into<iroha_error::Error>) -> PyErr {
-    PyException::new_err(err.into().report().to_string())
-}
 
 #[pymethods]
 impl KeyPair {
@@ -134,8 +25,17 @@ impl KeyPair {
     /// # Errors
     #[new]
     pub fn generate() -> PyResult<Self> {
-        let keys = iroha_crypto::KeyPair::generate().map_err(to_py_err)?;
-        Ok(Self { keys })
+        iroha_crypto::KeyPair::generate()
+            .map_err(to_py_err)
+            .map(Into::into)
+    }
+
+    #[staticmethod]
+    pub fn with_seed(seed: Vec<u8>) -> PyResult<Self> {
+        let cfg = KeyGenConfiguration::default().use_seed(seed);
+        iroha_crypto::KeyPair::generate_with_configuration(cfg)
+            .map_err(to_py_err)
+            .map(Into::into)
     }
 
     #[getter]
@@ -147,6 +47,18 @@ impl KeyPair {
     pub fn private(&self) -> Dict<PrivateKey> {
         Dict(self.private_key.clone())
     }
+}
+
+#[pyfunction]
+pub fn hash(bytes: Vec<u8>) -> Dict<Hash> {
+    Dict(Hash::new(&bytes))
+}
+
+#[pyfunction]
+pub fn sign(keys: KeyPair, payload: Vec<u8>) -> PyResult<Dict<Signature>> {
+    iroha_crypto::Signature::new(keys.into(), &payload)
+        .map_err(to_py_err)
+        .map(Dict)
 }
 
 #[pymethods]
@@ -163,7 +75,10 @@ impl Client {
         self.deref_mut()
             .request(query.into_inner())
             .map_err(to_py_err)
-            .map(Dict)
+            .map(|e| {
+                dbg!(&e);
+                Dict(e)
+            })
     }
 
     /// Sends transaction to peer
@@ -226,9 +141,9 @@ impl PyIterProtocol for EventIterator {
 
 #[rustfmt::skip]
 wrap_class!(
-    KeyPair        { keys: iroha_crypto::KeyPair }: Debug + Clone,
-    Client         { cl:   client::Client        }: Debug + Clone,
-    EventIterator  { it:   client::EventIterator }: Debug,
+    KeyPair        { keys: iroha_crypto::KeyPair   }: Debug + Clone,
+    Client         { cl:   client::Client          }: Debug + Clone,
+    EventIterator  { it:   client::EventIterator   }: Debug,
 );
 
 /// A Python module implemented in Rust.
